+++ protocols/identify/src/behaviour.rs
@@ -71,14 +71,6 @@ pub struct Config {
     ///
     /// Defaults to `rust-libp2p/<libp2p-identify-version>`.
     pub agent_version: String,
-    /// The initial delay before the first identification request
-    /// is sent to a remote on a newly established connection.
-    ///
-    /// Defaults to 0ms.
-    #[deprecated(note = "The `initial_delay` is no longer necessary and will be
-                completely removed since a remote should be able to instantly
-                answer to an identify request")]
-    pub initial_delay: Duration,
     /// The interval at which identification requests are sent to
     /// the remote on established connections after the first request,
     /// i.e. the delay between identification requests.
@@ -106,13 +98,11 @@ pub struct Config {
 impl Config {
     /// Creates a new configuration for the identify [`Behaviour`] that
     /// advertises the given protocol version and public key.
-    #[allow(deprecated)]
     pub fn new(protocol_version: String, local_public_key: PublicKey) -> Self {
         Self {
             protocol_version,
             agent_version: format!("rust-libp2p/{}", env!("CARGO_PKG_VERSION")),
             local_public_key,
-            initial_delay: Duration::from_millis(0),
             interval: Duration::from_secs(5 * 60),
             push_listen_addr_updates: false,
             cache_size: 100,
@@ -122,17 +112,6 @@ impl Config {
     /// Configures the agent version sent to peers.
     pub fn with_agent_version(mut self, v: String) -> Self {
         self.agent_version = v;
-        self
-    }
-
-    /// Configures the initial delay before the first identification
-    /// request is sent on a newly established connection to a peer.
-    #[deprecated(note = "The `initial_delay` is no longer necessary and will be
-                completely removed since a remote should be able to instantly
-                answer to an identify request thus also this setter will be removed")]
-    #[allow(deprecated)]
-    pub fn with_initial_delay(mut self, d: Duration) -> Self {
-        self.initial_delay = d;
         self
     }
 
@@ -235,7 +214,6 @@ impl NetworkBehaviour for Behaviour {
     type ConnectionHandler = Handler;
     type ToSwarm = Event;
 
-    #[allow(deprecated)]
     fn handle_established_inbound_connection(
         &mut self,
         _: ConnectionId,
@@ -244,7 +222,6 @@ impl NetworkBehaviour for Behaviour {
         remote_addr: &Multiaddr,
     ) -> Result<THandler<Self>, ConnectionDenied> {
         Ok(Handler::new(
-            self.config.initial_delay,
             self.config.interval,
             peer,
             self.config.local_public_key.clone(),
@@ -255,7 +232,6 @@ impl NetworkBehaviour for Behaviour {
         ))
     }
 
-    #[allow(deprecated)]
     fn handle_established_outbound_connection(
         &mut self,
         _: ConnectionId,
@@ -264,7 +240,6 @@ impl NetworkBehaviour for Behaviour {
         _: Endpoint,
     ) -> Result<THandler<Self>, ConnectionDenied> {
         Ok(Handler::new(
-            self.config.initial_delay,
             self.config.interval,
             peer,
             self.config.local_public_key.clone(),