+++ protocols/identify/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Add `Info` to the `libp2p-identify::Event::Pushed` to report pushed info.
   See [PR 4527](https://github.com/libp2p/rust-libp2p/pull/4527)
+- Remove deprecated `initial_delay`.
+  Identify requests are always sent instantly after the connection has been established.
+  See [PR 4735](https://github.com/libp2p/rust-libp2p/pull/4735)
 
 ## 0.43.1